+++ SwiftCSV/Parser.swift
@@ -21,12 +21,14 @@ extension CSV {
     ///   - rowLimit: Amount of rows to consume, beginning to count at `startAt`. Default value is `nil` to consume
     ///     the whole input string.
     ///   - rowCallback: Array of each row's columnar values, in order.
-    public func enumerateAsArray(startAt: Int = 0, rowLimit: Int? = nil, _ rowCallback: @escaping ([String]) -> ()) throws {
+    /// - Throws: `CSVParseError` or any error thrown by `rowCallback`
+    ///
+    public func enumerateAsArray(startAt: Int = 0, rowLimit: Int? = nil, _ rowCallback: @escaping ([String]) throws -> ()) throws {
 
         try Parser.enumerateAsArray(text: self.text, delimiter: self.delimiter, startAt: startAt, rowLimit: rowLimit, rowCallback: rowCallback)
     }
 
-    public func enumerateAsDict(_ block: @escaping ([String : String]) -> ()) throws {
+    public func enumerateAsDict(_ block: @escaping ([String : String]) throws -> ()) throws {
 
         try Parser.enumerateAsDict(header: self.header, content: self.text, delimiter: self.delimiter, block: block)
     }
@@ -55,12 +57,12 @@ enum Parser {
     ///   - rowLimit: Amount of rows to consume, beginning to count at `startAt`. Default value is `nil` to consume
     ///     the whole input string.
     ///   - rowCallback: Callback invoked for every parsed row between `startAt` and `limitTo` in `text`.
-    /// - Throws: `CSVParseError`
+    /// - Throws: `CSVParseError` or any error thrown by `rowCallback`
     static func enumerateAsArray(text: String,
                                  delimiter: CSVDelimiter,
                                  startAt offset: Int = 0,
                                  rowLimit: Int? = nil,
-                                 rowCallback: @escaping ([String]) -> ()) throws {
+                                 rowCallback: @escaping ([String]) throws -> ()) throws {
         let maxRowIndex = rowLimit.flatMap { $0 < 0 ? nil : offset + $0 }
 
         var currentIndex = text.startIndex
@@ -72,7 +74,7 @@ enum Parser {
 
         var rowIndex = 0
 
-        func finishRow() {
+        func finishRow() throws {
             defer {
                 rowIndex += 1
                 fields = []
@@ -81,7 +83,7 @@ enum Parser {
 
             guard rowIndex >= offset else { return }
             fields.append(String(field))
-            rowCallback(fields)
+            try rowCallback(fields)
         }
 
         var state: ParsingState = ParsingState(
@@ -118,12 +120,12 @@ enum Parser {
             }
 
             if !fields.isEmpty {
-                rowCallback(fields)
+                try rowCallback(fields)
             }
         }
     }
 
-    static func enumerateAsDict(header: [String], content: String, delimiter: CSVDelimiter, rowLimit: Int? = nil, block: @escaping ([String : String]) -> ()) throws {
+    static func enumerateAsDict(header: [String], content: String, delimiter: CSVDelimiter, rowLimit: Int? = nil, block: @escaping ([String : String]) throws -> ()) throws {
 
         let enumeratedHeader = header.enumerated()
 
@@ -133,7 +135,7 @@ enum Parser {
             for (index, head) in enumeratedHeader {
                 dict[head] = index < fields.count ? fields[index] : ""
             }
-            block(dict)
+            try block(dict)
         }
     }
 }